a bit of refactoring SettingsDataDonationViewModel
